better

+++ code/game/machinery/computer/security.dm
@@ -19,6 +19,7 @@
 	var/order = 1 // -1 = Descending - 1 = Ascending
 	light_color = LIGHT_COLOR_RED
 	var/can_change_job
+	var/can_set_arrest
 
 /obj/machinery/computer/secure_data/Initialize(mapload, obj/item/circuitboard/C)
 	. = ..()
@@ -469,6 +470,7 @@ What a mess.*/
 				active1 = null
 				active2 = null
 				can_change_job = null
+				can_set_arrest = null
 				playsound(src, 'sound/machines/terminal_off.ogg', 50, FALSE)
 
 			if("Log In")
@@ -480,6 +482,7 @@ What a mess.*/
 					authenticated = M.name
 					rank = JOB_NAME_AI
 					can_change_job = TRUE
+					can_set_arrest = TRUE
 					screen = 1
 				else if(IsAdminGhost(M))
 					active1 = null
@@ -487,16 +490,17 @@ What a mess.*/
 					authenticated = M.client.holder.admin_signature
 					rank = JOB_CENTCOM_CENTRAL_COMMAND
 					can_change_job = TRUE
+					can_set_arrest = TRUE
 					screen = 1
 				else if(I && check_access(I))
 					active1 = null
 					active2 = null
 					authenticated = I.registered_name
 					rank = I.assignment
-					for(var/each_access in list(ACCESS_CHANGE_IDS, ACCESS_CAPTAIN))
-						if(each_access in I.access)
-							can_change_job = TRUE
-							break
+					if(ACCESS_CHANGE_IDS in I.access)
+						can_change_job = TRUE
+					if(ACCESS_SEC_RECORDS in I.access)
+						can_set_arrest = TRUE
 					screen = 1
 				else
 					to_chat(usr, "<span class='danger'>Unauthorized Access.</span>")
@@ -758,7 +762,9 @@ What a mess.*/
 							if(istype(active2, /datum/data/record))
 								active2.fields["id"] = t1
 					if("fingerprint")
-						if(istype(active1, /datum/data/record))
+						if(!can_set_arrest)
+							alert(usr, "Your access is not sufficient to do this!")
+						else if(istype(active1, /datum/data/record))
 							var/t1 = stripped_input(usr, "Please input fingerprint hash:", "Secure. records", active1.fields["fingerprint"], null)
 							if(!canUseSecurityRecordsConsole(usr, t1, a1))
 								return
@@ -782,43 +788,51 @@ What a mess.*/
 								return
 							active1.fields["species"] = t1
 					if("upd_photo_front")
-						var/datum/picture/picture = get_picture(usr)
-						if(picture)
-							qdel(active1.fields["photo_front"])
-							//Lets center it to a 32x32.
-							var/icon/I = picture.picture_image
-							var/w = I.Width()
-							var/h = I.Height()
-							var/dw = w - 32
-							var/dh = w - 32
-							I.Crop(dw/2, dh/2, w - dw/2, h - dh/2)
-							var/obj/item/photo/new_photo = new
-							new_photo.picture = picture
-							active1.fields["photo_front"] = new_photo
+						if(!can_set_arrest)
+							alert(usr, "Your access is not sufficient to do this!")
+						else
+							var/datum/picture/picture = get_picture(usr)
+							if(picture)
+								qdel(active1.fields["photo_front"])
+								//Lets center it to a 32x32.
+								var/icon/I = picture.picture_image
+								var/w = I.Width()
+								var/h = I.Height()
+								var/dw = w - 32
+								var/dh = w - 32
+								I.Crop(dw/2, dh/2, w - dw/2, h - dh/2)
+								var/obj/item/photo/new_photo = new
+								new_photo.picture = picture
+								active1.fields["photo_front"] = new_photo
 					if("print_photo_front")
 						if(active1.fields["photo_front"])
 							if(istype(active1.fields["photo_front"], /datum/picture))
 								var/datum/picture/P = active1.fields["photo_front"]
 								print_photo(P.picture_image, active1.fields["name"])
 					if("upd_photo_side")
-						var/datum/picture/picture = get_picture(usr)
-						if(picture)
-							qdel(active1.fields["photo_side"])
-							//Lets center it to a 32x32.
-							var/icon/I = picture.picture_image
-							var/w = I.Width()
-							var/h = I.Height()
-							var/dw = w - 32
-							var/dh = w - 32
-							I.Crop(dw/2, dh/2, w - dw/2, h - dh/2)
-							active1.fields["photo_side"] = picture
+						if(!can_set_arrest)
+							alert(usr, "Your access is not sufficient to do this!")
+						else
+							var/datum/picture/picture = get_picture(usr)
+							if(picture)
+								qdel(active1.fields["photo_side"])
+								//Lets center it to a 32x32.
+								var/icon/I = picture.picture_image
+								var/w = I.Width()
+								var/h = I.Height()
+								var/dw = w - 32
+								var/dh = w - 32
+								I.Crop(dw/2, dh/2, w - dw/2, h - dh/2)
+								active1.fields["photo_side"] = picture
 					if("print_photo_side")
 						if(active1.fields["photo_side"])
 							if(istype(active1.fields["photo_side"], /datum/picture))
 								var/datum/picture/P = active1.fields["photo_side"]
 								print_photo(P.picture_image, active1.fields["name"])
 					if("crim_add")
-						if(istype(active1, /datum/data/record))
+						if(!can_set_arrest)
+							alert(usr, "Your access is not sufficient to do this!")
+						else if(istype(active1, /datum/data/record))
 							var/t1 = stripped_input(usr, "Please input crime name:", "Secure. records", "", null)
 							var/t2 = stripped_input(usr, "Please input crime details:", "Secure. records", "", null)
 							if(!canUseSecurityRecordsConsole(usr, t1, null, a2))
@@ -827,13 +841,17 @@ What a mess.*/
 							GLOB.data_core.addCrime(active1.fields["id"], crime)
 							usr.investigate_log(" has added a crime '<strong>[t1]</strong>': '[t2]' to [active1.fields["name"]] via security records.", INVESTIGATE_RECORDS)
 					if("crim_delete")
-						if(istype(active1, /datum/data/record))
+						if(!can_set_arrest)
+							alert(usr, "Your access is not sufficient to do this!")
+						else if(istype(active1, /datum/data/record))
 							if(href_list["cdataid"])
 								if(!canUseSecurityRecordsConsole(usr, "delete", null, a2))
 									return
 								GLOB.data_core.removeCrime(active1.fields["id"], href_list["cdataid"])
 					if("add_details")
-						if(istype(active1, /datum/data/record))
+						if(!can_set_arrest)
+							alert(usr, "Your access is not sufficient to do this!")
+						else if(istype(active1, /datum/data/record))
 							if(href_list["cdataid"])
 								var/t1 = stripped_input(usr, "Please input crime details:", "Secure. records", "", null)
 								if(!canUseSecurityRecordsConsole(usr, t1, null, a2))
@@ -841,7 +859,9 @@ What a mess.*/
 								GLOB.data_core.addCrimeDetails(active1.fields["id"], href_list["cdataid"], t1)
 								usr.investigate_log(" has set crime details ([t1]) to [active1.fields["name"]] via security records.", INVESTIGATE_RECORDS)
 					if("citation_add")
-						if(istype(active1, /datum/data/record))
+						if(!can_set_arrest)
+							alert(usr, "Your access is not sufficient to do this!")
+						else if(istype(active1, /datum/data/record))
 							var/maxFine = CONFIG_GET(number/maxfine)
 
 							var/t1 = stripped_input(usr, "Please input citation crime:", "Secure. records", "", null)
@@ -873,21 +893,27 @@ What a mess.*/
 							GLOB.data_core.addCitation(active1.fields["id"], crime)
 							usr.investigate_log(" has added a citation '<strong>[t1]</strong>' ([fine] credits) to [active1.fields["name"]] via security records.", INVESTIGATE_RECORDS)
 					if("citation_delete")
-						if(istype(active1, /datum/data/record))
+						if(!can_set_arrest)
+							alert(usr, "Your access is not sufficient to do this!")
+						else if(istype(active1, /datum/data/record))
 							if(href_list["cdataid"])
 								if(!canUseSecurityRecordsConsole(usr, "delete", null, a2))
 									return
 								usr.investigate_log(" has deleted [active1.fields["name"]]'s citation '[active1.fields["id"]]'.", INVESTIGATE_RECORDS)
 								GLOB.data_core.removeCitation(active1.fields["id"], href_list["cdataid"])
 					if("notes")
-						if(istype(active2, /datum/data/record))
+						if(!can_set_arrest)
+							alert(usr, "Your access is not sufficient to do this!")
+						else if(istype(active2, /datum/data/record))
 							var/t1 = stripped_input(usr, "Please summarize notes:", "Secure. records", active2.fields["notes"], null)
 							if(!canUseSecurityRecordsConsole(usr, t1, null, a2))
 								return
 							active2.fields["notes"] = t1
 							usr.investigate_log(" has set [active1.fields["name"]]'s security records important notes to '[t1]'.", INVESTIGATE_RECORDS)
 					if("criminal")
-						if(istype(active2, /datum/data/record))
+						if(!can_set_arrest)
+							alert(usr, "Your access is not sufficient to do this!")
+						else if(istype(active2, /datum/data/record) && )
 							temp = "<h5>Criminal Status:</h5>"
 							temp += "<ul>"
 							temp += "<li><a href='?src=[REF(src)];choice=Change Criminal Status;criminal2=none'>None</a></li>"
@@ -899,14 +925,14 @@ What a mess.*/
 							temp += "<li><a href='?src=[REF(src)];choice=Change Criminal Status;criminal2=released'>Discharged</a></li>"
 							temp += "</ul>"
 					if("rank")
-						if(can_change_job)
+						if(!can_change_job)
+							alert(usr, "Your access is not sufficient to do this!")
+						else if(istype(active1, /datum/data/record))
 							temp = "<h5>Rank:</h5>"
 							temp += "<ul>"
 							for(var/rank in get_all_jobs())
 								temp += "<li><a href='?src=[REF(src)];choice=Change Rank;rank=[rank]'>[rank]</a></li>"
 							temp += "</ul>"
-						else
-							alert(usr, "Your access is not sufficient to do this!")
 //TEMPORARY MENU FUNCTIONS
 			else//To properly clear as per clear screen.
 				temp=null
